Move Storage implementation for opendal::Operator to storage.rs
