+++ src/cache/storage.rs
@@ -10,7 +10,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::cache_io::{Cache, CacheMode, CacheWrite};
+use super::cache_io::{Cache, CacheMode, CacheRead, CacheWrite};
+#[cfg(any(
+    feature = "azure",
+    feature = "gcs",
+    feature = "gha",
+    feature = "memcached",
+    feature = "redis",
+    feature = "s3",
+    feature = "webdav",
+    feature = "oss",
+    feature = "cos"
+))]
+use crate::cache::utils::normalize_key;
 use crate::errors::*;
 use async_trait::async_trait;
 use std::time::Duration;
@@ -57,4 +69,105 @@ pub trait Storage: Send + Sync {
 
     /// Get the maximum storage size, if applicable.
     async fn max_size(&self) -> Result<Option<u64>>;
+}
+
+/// Implement storage for operator.
+#[cfg(any(
+    feature = "azure",
+    feature = "gcs",
+    feature = "gha",
+    feature = "memcached",
+    feature = "redis",
+    feature = "s3",
+    feature = "webdav",
+    feature = "oss",
+    feature = "cos"
+))]
+#[async_trait]
+impl Storage for opendal::Operator {
+    async fn get(&self, key: &str) -> Result<Cache> {
+        match self.read(&normalize_key(key)).await {
+            Ok(res) => {
+                let hit = CacheRead::from(std::io::Cursor::new(res.to_bytes()))?;
+                Ok(Cache::Hit(hit))
+            }
+            Err(e) if e.kind() == opendal::ErrorKind::NotFound => Ok(Cache::Miss),
+            Err(e) => {
+                warn!("Got unexpected error: {:?}", e);
+                Ok(Cache::Miss)
+            }
+        }
+    }
+
+    async fn put(&self, key: &str, entry: CacheWrite) -> Result<Duration> {
+        let start = std::time::Instant::now();
+
+        self.write(&normalize_key(key), entry.finish()?).await?;
+
+        Ok(start.elapsed())
+    }
+
+    async fn check(&self) -> Result<CacheMode> {
+        use opendal::ErrorKind;
+
+        let path = ".sccache_check";
+
+        // Read is required, return error directly if we can't read .
+        match self.read(path).await {
+            Ok(_) => (),
+            // Read not exist file with not found is ok.
+            Err(err) if err.kind() == ErrorKind::NotFound => (),
+            // Tricky Part.
+            //
+            // We tolerate rate limited here to make sccache keep running.
+            // For the worse case, we will miss all the cache.
+            //
+            // In some super rare cases, user could configure storage in wrong
+            // and hitting other services rate limit. There are few things we
+            // can do, so we will print our the error here to make users know
+            // about it.
+            Err(err) if err.kind() == ErrorKind::RateLimited => {
+                eprintln!("cache storage read check: {err:?}, but we decide to keep running");
+            }
+            Err(err) => bail!("cache storage failed to read: {:?}", err),
+        }
+
+        let can_write = match self.write(path, "Hello, World!").await {
+            Ok(_) => true,
+            Err(err) if err.kind() == ErrorKind::AlreadyExists => true,
+            // Tolerate all other write errors because we can do read at least.
+            Err(err) => {
+                eprintln!("storage write check failed: {err:?}");
+                false
+            }
+        };
+
+        let mode = if can_write {
+            CacheMode::ReadWrite
+        } else {
+            CacheMode::ReadOnly
+        };
+
+        debug!("storage check result: {mode:?}");
+
+        Ok(mode)
+    }
+
+    fn location(&self) -> String {
+        let meta = self.info();
+        format!(
+            "{}, name: {}, prefix: {}",
+            meta.scheme(),
+            meta.name(),
+            meta.root()
+        )
+    }
+
+    async fn current_size(&self) -> Result<Option<u64>> {
+        Ok(None)
+    }
+
+    async fn max_size(&self) -> Result<Option<u64>> {
+        Ok(None)
+    }
 }
